requirements.txt
django
djangorestframework
psycopg2-binary
pytest-django
model_bakery
pytest-cov